+++ tencentcloud/apm/v20210622/errors.go
@@ -26,6 +26,15 @@ const (
 	// 操作失败。
 	FAILEDOPERATION = "FailedOperation"
 
+	// appid和实例信息不匹配。
+	FAILEDOPERATION_APPIDNOTMATCHINSTANCEINFO = "FailedOperation.AppIdNotMatchInstanceInfo"
+
+	// 实例ID为空。
+	FAILEDOPERATION_INSTANCEIDISEMPTY = "FailedOperation.InstanceIdIsEmpty"
+
+	// apm实例不存在。
+	FAILEDOPERATION_INSTANCENOTFOUND = "FailedOperation.InstanceNotFound"
+
 	// 非内网vpc。
 	FAILEDOPERATION_NOTINNERVPC = "FailedOperation.NotInnerVPC"