+++ tencentcloud/clb/v20180317/models.go
@@ -3787,7 +3787,7 @@ type HealthCheck struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	HttpVersion *string `json:"HttpVersion,omitempty" name:"HttpVersion"`
 
-	// 自定义探测相关参数。健康检查原IP类型：0（使用LB的VIP做为源IP），1（使用100.64网段IP做为源IP），默认值：0
+	// 自定义探测相关参数。健康检查原IP类型：0（使用LB的VIP作为源IP），1（使用100.64网段IP作为源IP），默认值：0
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	SourceIpType *int64 `json:"SourceIpType,omitempty" name:"SourceIpType"`
 }